requirements.txt
streamlit
pandas
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
openai
pypdf2
